conduccion/requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.3

# Configuration
django-environ==0.4.1
whitenoise==3.2.2


# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.29.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.6.0
redis>=2.10.5





# Your custom requirements go here
djangorestframework==3.5.3
markdown==2.6.6
django-jet==1.0.2
django-cors-headers==1.2.0
django-admin-sortable2==0.6.4
